% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPO.R, R/NPO_allData.R
\name{NPO}
\alias{NPO}
\title{Run the Network Parameter Outlier (NPO) algorithm.}
\usage{
NPO(...)

NPO(...)
}
\description{
Run the Network Parameter Outlier (NPO) algorithm.

Run the Network Parameter Outlier (NPO) algorithm.
}
\examples{
\dontrun{
  First time the function is called:
  NPO(dbName='NSME_halo',path='NPO/Analysis/NPO/tests/testData/Halo/11',data='rodentMSO',institution='Yale',lab='NSME',experiment='Halo10sec_10x',subject=11,signalType='AP',centerTime=0,iterationType='directory',range=c(-3000,2000), '--restart' )

  Subsequent calls:
  NPO()
}
\dontrun{
  First time the function is called:
  NPO(dbName='NSME_halo',path='NPO/Analysis/NPO/tests/testData/Halo/11',data='rodentMSO',institution='Yale',lab='NSME',experiment='Halo10sec_10x',subject=11,signalType='AP',centerTime=0,iterationType='directory',range=c(-3000,2000), '--restart' )

  Subsequent calls:
  NPO()
}
}
